refactor: create service package and separate CouchDB functionality
- Create cmd/service/ package for better organization
- Move CouchDB commands to cmd/service/couchdb.go
- Simplify cmd/service.go to use the new service package
- Improve code maintainability and scalability for future services
